+++ docs/example/element-plus/collapse-button/position.vue
@@ -1,40 +1,46 @@
 <template>
   <div class="flex">
-    <div class="collapse-contain border-radius flex-center">
-      <div v-show="!isShow">
-        点击左侧按钮显示内容
+    <div class="flex-shrink">
+      <div class="collapse-contain border-radius flex-center">
+        <div v-show="!isShow">
+          点击左侧按钮显示内容
+        </div>
+        <ul class="content" v-show="isShow">
+          <li>我见众生皆草木</li>
+          <li>唯有见你是青山</li>
+          <li>愿有岁月可回首</li>
+          <li>且以你我共白头</li>
+        </ul>
+        <k-collapse-button position="left" @click="isShow = !isShow" />
       </div>
-      <ul class="content" v-show="isShow">
-        <li>我见众生皆草木</li>
-        <li>唯有见你是青山</li>
-        <li>愿有岁月可回首</li>
-        <li>且以你我共白头</li>
-      </ul>
-      <k-collapse-button position="left" @click="isShow = !isShow" />
     </div>
-    <div class="collapse-contain border-radius flex-center">
-      <div v-show="!isShow1">
-        点击下侧按钮显示内容
+    <div class="flex-shrink">
+      <div class="collapse-contain border-radius flex-center">
+        <div v-show="!isShow1">
+          点击下侧按钮显示内容
+        </div>
+        <ul class="content" v-show="isShow1">
+          <li>我见众生皆草木</li>
+          <li>唯有见你是青山</li>
+          <li>愿有岁月可回首</li>
+          <li>且以你我共白头</li>
+        </ul>
+        <k-collapse-button position="bottom" @click="isShow1 = !isShow1" />
       </div>
-      <ul class="content" v-show="isShow1">
-        <li>我见众生皆草木</li>
-        <li>唯有见你是青山</li>
-        <li>愿有岁月可回首</li>
-        <li>且以你我共白头</li>
-      </ul>
-      <k-collapse-button position="bottom" @click="isShow1 = !isShow1" />
     </div>
-    <div class="collapse-contain border-radius flex-center">
-      <div v-show="!isShow2">
-        点击右侧按钮显示内容
+    <div class="flex-shrink">
+      <div class="collapse-contain border-radius flex-center">
+        <div v-show="!isShow2">
+          点击右侧按钮显示内容
+        </div>
+        <ul class="content" v-show="isShow2">
+          <li>我见众生皆草木</li>
+          <li>唯有见你是青山</li>
+          <li>愿有岁月可回首</li>
+          <li>且以你我共白头</li>
+        </ul>
+        <k-collapse-button position="right" :style="{top:'10px'}" @click="isShow2 = !isShow2" />
       </div>
-      <ul class="content" v-show="isShow2">
-        <li>我见众生皆草木</li>
-        <li>唯有见你是青山</li>
-        <li>愿有岁月可回首</li>
-        <li>且以你我共白头</li>
-      </ul>
-      <k-collapse-button position="right" :style="{top:'10px'}" @click="isShow2 = !isShow2" />
     </div>
   </div>
 </template>